+++ tests/testthat/test-tar_option_set.R
@@ -153,11 +153,11 @@ tar_test("deprecated error = \"workspace\"", {
 })
 
 tar_test("memory", {
-  expect_equal(tar_option_get("memory"), "auto")
-  tar_option_set(memory = "transient")
   expect_equal(tar_option_get("memory"), "transient")
+  tar_option_set(memory = "persistent")
+  expect_equal(tar_option_get("memory"), "persistent")
   tar_option_reset()
-  expect_equal(tar_option_get("memory"), "auto")
+  expect_equal(tar_option_get("memory"), "transient")
   expect_error(
     tar_option_set(memory = "invalid"),
     class = "tar_condition_validate"
@@ -255,11 +255,11 @@ tar_test("resources", {
 })
 
 tar_test("storage", {
-  expect_equal(tar_option_get("storage"), "main")
-  tar_option_set(storage = "worker")
   expect_equal(tar_option_get("storage"), "worker")
-  tar_option_reset()
+  tar_option_set(storage = "main")
   expect_equal(tar_option_get("storage"), "main")
+  tar_option_reset()
+  expect_equal(tar_option_get("storage"), "worker")
   expect_error(
     tar_option_set(storage = "invalid"),
     class = "tar_condition_validate"
@@ -267,11 +267,11 @@ tar_test("storage", {
 })
 
 tar_test("retrieval", {
-  expect_equal(tar_option_get("retrieval"), "main")
-  tar_option_set(retrieval = "worker")
   expect_equal(tar_option_get("retrieval"), "worker")
-  tar_option_reset()
+  tar_option_set(retrieval = "main")
   expect_equal(tar_option_get("retrieval"), "main")
+  tar_option_reset()
+  expect_equal(tar_option_get("retrieval"), "worker")
   expect_error(
     tar_option_set(retrieval = "invalid"),
     class = "tar_condition_validate"